+++ packages/editor/editor-core/src/__tests__/visual-regression/common/layouts.ts
@@ -1,10 +1,18 @@
 import { initFullPageEditorWithAdf, snapshot, Device } from '../_utils';
-import { selectors } from '../../__helpers/page-objects/_editor';
 import * as col2 from './__fixtures__/column2-adf.json';
 import * as col3 from './__fixtures__/column3-adf.json';
+import * as colLeftSidebar from './__fixtures__/columnLeftSidebar-adf.json';
+import * as columnRightSidebar from './__fixtures__/columnRightSidebar-adf.json';
 
 describe('Layouts:', () => {
   let page: any;
+
+  const layouts = [
+    { name: '2 columns', adf: col2 },
+    { name: '3 columns', adf: col3 },
+    { name: 'left sidebar', adf: colLeftSidebar },
+    { name: 'right sidebar', adf: columnRightSidebar },
+  ];
 
   beforeAll(async () => {
     // @ts-ignore
@@ -15,31 +23,24 @@ describe('Layouts:', () => {
     await snapshot(page, 0.02);
   });
 
-  describe('2 columns', () => {
-    it('should correctly render layout on MDPI', async () => {
-      await initFullPageEditorWithAdf(page, col2, Device.LaptopMDPI);
-    });
+  layouts.forEach(layout => {
+    describe(layout.name, () => {
+      const initEditor = async (device: Device) =>
+        initFullPageEditorWithAdf(page, layout.adf, device, undefined, {
+          allowLayouts: { allowBreakout: true, UNSAFE_addSidebarLayouts: true },
+        });
 
-    it('should stack layout on smaller ipad', async () => {
-      await initFullPageEditorWithAdf(page, col2, Device.iPad);
-      await page.click(selectors.layoutDataSection);
-    });
+      it('should correctly render layout on MDPI', async () => {
+        await initEditor(Device.LaptopMDPI);
+      });
 
-    it('should stack layout on smaller iPhone', async () => {
-      await initFullPageEditorWithAdf(page, col2, Device.iPhonePlus);
-      await page.click(selectors.layoutDataSection);
-    });
-  });
+      it('should stack layout on smaller ipad', async () => {
+        await initEditor(Device.iPad);
+      });
 
-  describe('3 columns', () => {
-    it('should correctly render layout', async () => {
-      await initFullPageEditorWithAdf(page, col3, Device.LaptopMDPI);
-      await page.click(selectors.layoutDataSection);
-    });
-
-    it('should stack layout on smaller screen sizes', async () => {
-      await initFullPageEditorWithAdf(page, col3, Device.iPhonePlus);
-      await page.click(selectors.layoutDataSection);
+      it('should stack layout on smaller iPhone', async () => {
+        await initEditor(Device.iPhonePlus);
+      });
     });
   });
 });