+++ packages/elements/share/src/components/ShareDialogWithTrigger.tsx
@@ -14,9 +14,11 @@ import { messages } from '../i18n';
 import {
   ConfigResponse,
   DialogContentState,
+  DialogPlacement,
   Flag,
   OriginTracing,
   ShareButtonStyle,
+  RenderCustomTriggerButton,
   ADMIN_NOTIFIED,
   OBJECT_SHARED,
 } from '../types';
@@ -29,10 +31,6 @@ import {
 } from './analytics';
 import { ShareButton } from './ShareButton';
 import { ShareForm } from './ShareForm';
-
-type RenderChildren = (
-  args: { onClick: () => void; loading?: boolean; error?: ShareError },
-) => React.ReactNode;
 
 type DialogState = {
   isDialogOpen: boolean;
@@ -50,9 +48,9 @@ type ShareError = {
 
 export type Props = {
   config?: ConfigResponse;
-  children?: RenderChildren;
+  children?: RenderCustomTriggerButton;
   copyLink: string;
-  dialogPlacement?: string;
+  dialogPlacement?: DialogPlacement;
   isDisabled?: boolean;
   loadUserOptions?: LoadOptions;
   onLinkCopy?: Function;
@@ -85,8 +83,8 @@ class ShareDialogWithTriggerInternal extends React.Component<
 > {
   static defaultProps = {
     isDisabled: false,
-    dialogPlacement: 'bottom-end',
-    shouldCloseOnEscapePress: false,
+    dialogPlacement: 'bottom-end' as 'bottom-end',
+    shouldCloseOnEscapePress: true,
     triggerButtonAppearance: 'subtle' as 'subtle',
     triggerButtonStyle: 'icon-only' as 'icon-only',
   };
@@ -149,7 +147,8 @@ class ShareDialogWithTriggerInternal extends React.Component<
 
   private handleKeyDown = (event: React.KeyboardEvent<HTMLDivElement>) => {
     const { isDialogOpen } = this.state;
-    if (isDialogOpen) {
+    const { shouldCloseOnEscapePress } = this.props;
+    if (isDialogOpen && shouldCloseOnEscapePress) {
       switch (event.key) {
         case 'Escape':
           event.stopPropagation();