+++ packages/media/media-viewer/src/newgen/item-viewer.tsx
@@ -107,6 +107,20 @@ export class ItemViewerBase extends React.Component<Props, State> {
     });
   };
 
+  private onCanPlay = (fileState: FileState) => () => {
+    if (fileState.status === 'processed') {
+      this.fireAnalytics(mediaFileLoadSucceededEvent(fileState));
+    }
+  };
+
+  private onError = (fileState: FileState) => () => {
+    if (fileState.status === 'processed') {
+      this.fireAnalytics(
+        mediaFileLoadFailedEvent(fileState.id, 'Playback failed', fileState),
+      );
+    }
+  };
+
   private renderFileState(item: FileState) {
     if (item.status === 'error') {
       return this.renderError('previewFailed', item);
@@ -139,6 +153,8 @@ export class ItemViewerBase extends React.Component<Props, State> {
           <AudioViewer
             showControls={showControls}
             featureFlags={featureFlags}
+            onCanPlay={this.onCanPlay(item)}
+            onError={this.onError(item)}
             {...viewerProps}
           />
         );
@@ -147,6 +163,8 @@ export class ItemViewerBase extends React.Component<Props, State> {
           <VideoViewer
             showControls={showControls}
             featureFlags={featureFlags}
+            onCanPlay={this.onCanPlay(item)}
+            onError={this.onError(item)}
             {...viewerProps}
           />
         );