+++ packages/core/button/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @atlaskit/button
 
+## 12.0.1
+- Updated dependencies [d263485853](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/d263485853):
+  - @atlaskit/spinner@10.0.6
+
 ## 12.0.0
 - [major] [1e826b2966](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/1e826b2966):